requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
dj-database-url==0.5.0
Django==3.0
django-colorful==1.3
django-countries==5.5
django-crispy-forms==1.8.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==5.0.1
wrapt==1.11.2
